requirements.txt
Flask
cassandra-driver==3.24.0